requirements.txt
anyio==3.6.1
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
python-telegram-bot==v20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio
tornado==6.2
tzdata==2022.1
tzlocal==4.2